+++ test/aead_on_off.cpp
@@ -72,7 +72,7 @@ void aead_encrypt_test(
     cout << endl;
 
     cout << "ctxt: ";
-    for (int i = 0; i < msg_len; i++) {
+    for (size_t i = 0; i < msg_len; i++) {
         cout << hex << (int)ctxt[i];
     }
     cout << endl;
@@ -172,7 +172,7 @@ void aead_decrypt_test(
     if (party == ALICE) {
         cout << "ALICE res: " << res << endl;
         cout << "ALICE msg: ";
-        for (int i = 0; i < msg_len; i++) {
+        for (size_t i = 0; i < msg_len; i++) {
             cout << hex << (int)msg[i];
         }
         cout << endl;
@@ -180,7 +180,7 @@ void aead_decrypt_test(
     } else {
         cout << "BOB res: " << res << endl;
         cout << "BOB msg: ";
-        for (int i = 0; i < msg_len; i++) {
+        for (size_t i = 0; i < msg_len; i++) {
             cout << hex << (int)msg[i];
         }
         cout << endl;
