+++ test/handshake_on_off.cpp
@@ -109,11 +109,11 @@ void handshake_test(
     bool res = aead_c_server->decrypt(io, msg, ctxt, finished_msg_length, tag, aad, aad_len,
                                       iv_c_oct, 8, party);
     cout << "res: " << res << endl;
-    for (int i = 0; i < finished_msg_length; i++)
+    for (size_t i = 0; i < finished_msg_length; i++)
         cout << hex << (int)msg[i];
     cout << endl;
 
-    for (int i = 0; i < finished_msg_length; i++)
+    for (size_t i = 0; i < finished_msg_length; i++)
         cout << hex << (int)hs->client_ufin[i];
     cout << endl;
