+++ test/post_record.cpp
@@ -87,11 +87,11 @@ void post_record_test(IO* io, IO* io_opt, COT<IO>* cot, int party) {
     bool res = aead_c_server->decrypt(io, msg, finc_ctxt, finished_msg_length, finc_tag, aad,
                                       aad_len, iv_c_oct, 8, party);
     cout << "res: " << res << endl;
-    for (int i = 0; i < finished_msg_length; i++)
+    for (size_t i = 0; i < finished_msg_length; i++)
         cout << hex << (int)msg[i];
     cout << endl;
 
-    for (int i = 0; i < finished_msg_length; i++)
+    for (size_t i = 0; i < finished_msg_length; i++)
         cout << hex << (int)hs->client_ufin[i];
     cout << endl;
